Move custom render config
